requirements.txt
langchain
langchain_openai
langchain_community
sentence_transformers
chromadb
Gitpython
tiktoken
python-dotenv
flask
-e .